[RISCV][GISel] Add manual isel for s8/s16/s32 load/store for the GPR bank.
GISel doesn't distinquish integer and FP loads and stores. We only
know which it is after register bank selection. This results in
s16/s32 loads/stores on the GPR register bank that needs to be
selected. This required extra isel patterns not needed for SDAG
and adding i16 and i32 to the GPR register class. Having i16/i32
on the GPR register class makes type interfence in tablegen less
effective, requiring explicit casts to be added to patterns.

This patch removes the extra isel patterns and replaces it with custom
instruction selection similar to what is done on AArch64.