[RISCV] Fix copy/paste mistake in Sh3Add_UWPat.
This pattern appears to be a copy of the pattern in Sh3AddPat
but using sh3add.uw instead of sh3add. I think this is mistake and
the pattern should be the equivalent of the first pattern from
Sh1Add_UWPat and Sh2Add_UWPat.

These classes were created to share with Andes in a788a1abd9c88, but
there was so many test changes in there that we must have overlooked the
changes to Zba codegen.